+++ changelog.md
@@ -11,8 +11,15 @@ date - post `commit ID (short version)`
 - Single Drive Command File for robot-rel support 
 with trigger for more percise velocity control
 
+## 2026/2/03 - post `22a624b`
+- update Changelog
+- Remove State Machine
+
+## 2026/2/03 - post `f8a40ea`
+- codes for secondry shooter (!)
+
 ## 2026/2/02 - post `39002e2`
-ShooterIOKraken (!)
+- ShooterIOKraken (!)
 
 ## 2026/1/30 - post `80238d7`
 Subsystem Draft